Moved broken thread to owner controller. Also removed unneccessarry method from view bit - it is comtaind in hte modl.
