+++ src/controller/UploadTabController.java
@@ -18,6 +18,7 @@ import util.ExperimentData;
 import util.GenomeReleaseData;
 
 import gui.ErrorDialog;
+import gui.UploadFileRow;
 import gui.UploadTab;
 import gui.UploadToExistingExpPanel;
 import model.ErrorLogger;
@@ -46,11 +47,13 @@ public class UploadTabController {
                 SelectFilesToNewExpListener());
         uploadTab.getNewExpPanel().addUploadButtonListener(
                 UploadNewExpListener());
-        uploadTab.setOngoingUploads(model.getOngoingUploads());
         uploadTab.getNewExpPanel().addUploadSelectedFilesListener(
                 UploadSelectedFilesListener());
         uploadTab.getNewExpPanel().addSpeciesSelectedListener(
                 SpeciesSelectedListener());
+
+
+        updateProgress();
     }
 
     /**
@@ -437,4 +440,49 @@ public class UploadTabController {
         };
     }
 
+    /**
+     * Method updating the progress of ongoing uploads.
+     */
+    private void updateProgress() {
+        new Thread(new Runnable() {
+            private boolean running;
+
+            @Override
+            public void run() {
+                running = true;
+                while (running) {
+                    for (File key : uploadTab.getNewExpPanel().getFileRows().keySet()) {
+                        UploadFileRow row = uploadTab.getNewExpPanel().getFileRows()
+                                .get(key);
+                        for (HTTPURLUpload upload : model.getOngoingUploads()) {
+                            if (upload.getFileName().equals(row.getFileName())) {
+                                row.updateProgressBar(upload
+                                        .getCurrentProgress());
+                            }
+                        }
+                    }
+                    for (File key : uploadTab.getExistExpPanel().getFileRows()
+                            .keySet()) {
+                        UploadFileRow row = uploadTab.getExistExpPanel()
+                                .getFileRows().get(key);
+                        for (HTTPURLUpload upload : model.getOngoingUploads()) {
+                            if (upload.getFileName().equals(row.getFileName())) {
+                                row.updateProgressBar(upload
+                                        .getCurrentProgress());
+                            }
+                        }
+                    }
+                    try {
+                        Thread.sleep(100);
+                    } catch (InterruptedException e) {
+                        ErrorLogger.log(e);
+                        running = false;
+                    }
+                    // TODO: THIS IS BROKEN, more is created on each logout-in
+                    // !!!
+//                    System.err.println(this.toString());
+                }
+            }
+        }).start();
+    }
 }
