+++ src/gui/GUI.java
@@ -717,10 +717,6 @@ public class GUI extends JFrame implements GenomizerView {
         processTab.showProcessFeedback(processFeedbackData);
     }
 
-    public void setOngoingUploads(
-            CopyOnWriteArrayList<HTTPURLUpload> ongoingUploads) {
-        uploadTab.setOngoingUploads(ongoingUploads);
-    }
 
     /**
      * returns the RatioCalcPopup
